Move the ServerClient code into a dedicated file
